requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx
openai
cryptography
PyJWT
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client